requirements.txt
openai==0.28
streamlit
python-decouple
